Fixing flow.

+++ ProceduralLandscapeGeneration/ErosionSimulation/HydraulicErosion/Grid/Shaders/4VerticalSuspendDepositeComputeShader.glsl
@@ -179,7 +179,7 @@ float SuspendFromLayerTop(uint index, uint layer, float requiredSediment)
     }
     for(int rockType = int(mapGenerationConfiguration.RockTypeCount) - 1; rockType >= 0; rockType--)
     {
-        uint offsetIndex = index + HeightMapRockTypeOffset(rockType) + HeightMapLayerOffset(layer);
+        uint offsetIndex = index + HeightMapRockTypeOffset(uint(rockType)) + HeightMapLayerOffset(layer);
         float height = heightMap[offsetIndex];
         float hardness = (1.0 - rockTypesConfiguration[rockType].Hardness);
         float toBeSuspendedSediment = requiredSediment * hardness;
